pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    { name = "Sergei Bykov", email = "ss.bykoff@gmail.com" }
]
packages = [
    { include = "app" },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "fastapi (>=0.115.7,<0.116.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi-users[sqlalchemy] (>=14.0.1,<15.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "sqladmin (>=0.20.1,<0.21.0)",
    "orjson (>=3.10.15,<4.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "google-api-python-client (>=2.163.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)",
    "aiohttp (>=3.11.14,<4.0.0)",
    "types-requests (>=2.32.0.20250328,<3.0.0.0)",
    "google-api-python-client-stubs (>=1.29.0,<2.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
black = "^25.1.0"

